new branch

+++ GameLogicBlackJack.BusinessLogic/Interface/IBalance.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using GameLogicBlackJack.Models;
 
 namespace GameLogicBlackJack.BusinessLogic.Interface
 {
     public interface IBalance
     {
-        Decimal Balance { get; set; }
+        Int32 Balance { get; set; }
     }
 }
