+++ GameLogicBlackJack.BusinessLogic/Interface/IBet.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using GameLogicBlackJack.Models;
 
 namespace GameLogicBlackJack.BusinessLogic.Interface
 {
     public interface IBet
     {
-        Decimal Bet { get; set; }
+        Int32 Bet { get; set; }
     }
 }
